+++ include/engine/graph/RenderGraphTypes.h
@@ -60,7 +60,8 @@ namespace engine::graph {
     RasterDepthTestCount,
     RasterDepthPassCount,
     RasterShadeCount,
-    RasterColorWriteCount
+    RasterColorWriteCount,
+    HybridVisibility
   };
 
   /**